cmds/exp/disk_unlock: Split out HSS key generation
The key generation functionality used for the disk_unlock command will
also be used to generate keys for NVMe disk unlocking.  This commit
splits that functionality into a separate package hsskey.
