+++ cmds/exp/disk_unlock/disk_unlock.go
@@ -1,42 +1,26 @@
-// Copyright 2020 the u-root Authors. All rights reserved
+// Copyright 2022 the u-root Authors. All rights reserved
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// The disk_unlock command is used to unlock a disk drive as follows:
-//  1. Via BMC, read a 32-byte secret seed known as the Host Secret Seed (HSS)
-//     using the OpenBMC IPMI blob transfer protocol
-//  2. Compute a password as follows:
-//     We get the deterministically computed 32-byte HDKF-SHA256 using:
-//     - salt: "SKM PROD_V2 ACCESS"
-//     - hss: 32-byte HSS
-//     - device identity: strings formed by concatenating the assembly serial
-//     number, the _ character, and the assembly part number.
-//  3. Unlock the drive with the given password
-//  4. Update the partition table for the disk
+// The disk_unlock command is used to unlock a disk drive with a
+// HSS-derived password, and rescan the drive to enumerate the
+// unlocked partitions.
 package main
 
 import (
-	"crypto/sha256"
 	"flag"
 	"fmt"
 	"io"
 	"log"
 	"os"
 	"path/filepath"
-	"strings"
 
-	"github.com/u-root/u-root/pkg/ipmi"
-	"github.com/u-root/u-root/pkg/ipmi/blobs"
+	"github.com/u-root/u-root/pkg/hsskey"
 	"github.com/u-root/u-root/pkg/mount/block"
 	"github.com/u-root/u-root/pkg/mount/scuzz"
-	"golang.org/x/crypto/hkdf"
 )
 
 const (
-	hostSecretSeedLen = 32
-
-	passwordSalt = "SKM PROD_V2 ACCESS"
-
 	// Master Password ID for SKM-based unlock.
 	skmMPI = 0x0601
 )
@@ -53,97 +37,6 @@ func verboseLog(msg string) {
 	if *verbose {
 		log.Print(msg)
 	}
-}
-
-// readHssBlob reads a host secret seed from the given blob id.
-func readHssBlob(id string, h *blobs.BlobHandler) (data []uint8, rerr error) {
-	sessionID, err := h.BlobOpen(id, blobs.BMC_BLOB_OPEN_FLAG_READ)
-	if err != nil {
-		return nil, fmt.Errorf("IPMI BlobOpen for %s failed: %v", id, err)
-	}
-	defer func() {
-		// If the function returned successfully but failed to close the blob,
-		// return an error.
-		if err := h.BlobClose(sessionID); err != nil && rerr == nil {
-			rerr = fmt.Errorf("IPMI BlobClose %s failed: %v", id, err)
-		}
-	}()
-
-	data, err = h.BlobRead(sessionID, 0, hostSecretSeedLen)
-	if err != nil {
-		return nil, fmt.Errorf("IPMI BlobRead %s failed: %v", id, err)
-	}
-
-	if len(data) != hostSecretSeedLen {
-		return nil, fmt.Errorf("HSS size incorrect: got %d for %s", len(data), id)
-	}
-
-	return data, nil
-}
-
-// getAllHss reads all host secret seeds over IPMI.
-func getAllHss() ([][]uint8, error) {
-	i, err := ipmi.Open(0)
-	if err != nil {
-		return nil, err
-	}
-	h := blobs.NewBlobHandler(i)
-
-	blobCount, err := h.BlobGetCount()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get blob count: %v", err)
-	}
-
-	hssList := [][]uint8{}
-	seen := make(map[string]bool)
-	skmSubstr := "/skm/hss/"
-
-	// Read from all */skm/hss/* blobs.
-	for j := 0; j < blobCount; j++ {
-		id, err := h.BlobEnumerate(j)
-		if err != nil {
-			return nil, fmt.Errorf("failed to enumerate blob %d: %v", j, err)
-		}
-
-		if !strings.Contains(id, skmSubstr) {
-			continue
-		}
-
-		hss, err := readHssBlob(id, h)
-		if err != nil {
-			log.Printf("failed to read HSS of id %s: %v", id, err)
-			continue
-		}
-
-		msg := fmt.Sprintf("HSS Entry: Id=%s", id)
-		if *verboseNoSanitize {
-			msg = msg + fmt.Sprintf(",Seed=%x", hss)
-		}
-		verboseLog(msg)
-
-		hssStr := fmt.Sprint(hss)
-		if _, ok := seen[hssStr]; !ok {
-			seen[hssStr] = true
-			hssList = append(hssList, hss)
-		}
-	}
-
-	return hssList, nil
-}
-
-// Compute the password deterministically as the 32-byte HDKF-SHA256 of the
-// HSS plus the device identity.
-func genPassword(hss []byte, info *scuzz.Info) ([]byte, error) {
-	hash := sha256.New
-	devID := fmt.Sprintf("%s_%s", info.Serial, info.Model)
-
-	r := hkdf.New(hash, hss, ([]byte)(passwordSalt), ([]byte)(devID))
-	key := make([]byte, 32)
-
-	if _, err := io.ReadFull(r, key); err != nil {
-		return nil, err
-	}
-	return key, nil
 }
 
 // writeFile is ioutil.WriteFile but disallows creating new file
@@ -166,18 +59,6 @@ func writeFile(filename string, contents string) error {
 func main() {
 	flag.Parse()
 
-	// Obtain 32 byte Host Secret Seed (HSS) from IPMI.
-	hssList, err := getAllHss()
-	if err != nil {
-		log.Fatalf("error getting HSS: %v", err)
-	}
-
-	if len(hssList) == 0 {
-		log.Fatalf("no HSS found - can't unlock disk.")
-	}
-
-	verboseLog(fmt.Sprintf("Found %d Host Secret Seeds.", len(hssList)))
-
 	// Open the disk. Read its identity, and use it to unlock the disk.
 	sgdisk, err := scuzz.NewSGDisk(*disk)
 	if err != nil {
@@ -190,6 +71,18 @@ func main() {
 	}
 
 	verboseLog(fmt.Sprintf("Disk info for %s: %s", *disk, info.String()))
+
+	// Obtain 32 byte Host Secret Seed (HSS) from IPMI.
+	hssList, err := hsskey.GetAllHss(*verbose, *verboseNoSanitize)
+	if err != nil {
+		log.Fatalf("error getting HSS: %v", err)
+	}
+
+	if len(hssList) == 0 {
+		log.Fatalf("no HSS found - can't unlock disk.")
+	}
+
+	verboseLog(fmt.Sprintf("Found %d Host Secret Seeds.", len(hssList)))
 
 	switch {
 	case !info.SecurityStatus.SecurityEnabled():
@@ -219,7 +112,7 @@ func main() {
 
 TryAllHSS:
 	for i, hss := range hssList {
-		key, err := genPassword(hss, info)
+		key, err := hsskey.GenPassword(hss, ([]byte)(info.Serial), ([]byte)(info.Model))
 		if err != nil {
 			log.Printf("Couldn't generate password with HSS %d: %v", i, err)
 			continue
